# Ctrl+Young_4/Biblioteca-Digital-Com-MongoDB/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Biblioteca Digital API",
        "description": "Documentação da API da Biblioteca Digital"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
        "/books": {
            "post": {
                "summary": "Criar um novo livro",
                "description": "Adiciona um novo livro à biblioteca digital.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description:": "Dados do livro a ser criado",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "O Senhor dos Anéis"
                                },
                                "author": {
                                    "type": "string",
                                    "example": "J.R.R. Tolkien"
                                },
                                "year": {
                                    "type": "integer",
                                    "example": 1954
                                },
                                 "genre": {
                                    "type": "string",
                                    "example": "Fantasia"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "get": {
                "summary": "Listar livros",
                "description": "Lista os livros cadastrados no banco de dados da biblioteca digital.",
                "responses": {
                    "200": {
                        "description": "Uma lista de livros"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/books/{bookId}": {
            "get": {
                "summary": "Obter um livro pelo ID",
                "description": "Lista os detalhes de um livro específico pelo seu ID",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID do livro a ser retornado",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do livro"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "patch": {
                "summary": "Atualizar dados do livro",
                "description": "Atualiza os dados informados de um livro pelo seu ID",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID do livro a alterar os dados",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Dados parciais do livro a serem alterados (informar apenas os campos a serem modificados)",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "O Senhor dos Anéis"
                                },
                                "author": {
                                    "type": "string",
                                    "example": "J.R.R. Tolkien"
                                },
                                "year": {
                                    "type": "integer",
                                    "example": 1954
                                },
                                 "genre": {
                                    "type": "string",
                                    "example": "Fantasia"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do livro após a alteração"
                    },
                    "500": {
                        "description": "Erro ao atualizar o livro"
                    }
                }
            },
            "delete": {
                "summary": "Deletar um livro",
                "description": "Excluir um livro da biblioteca pelo seu ID",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Id do livro a ser deletado",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro excluído com sucesso"
                    },
                    "500": {
                        "description": "Erro ao excluir o livro"
                    }
                }
            }
        }
    }
}